Separate print in another class
