+++ srp/src/main/java/com/codurance/srp/Printer.java
@@ -1,0 +1,13 @@
+package com.codurance.srp;
+
+public class Printer {
+    private final Console console;
+
+    public Printer(Console console) {
+        this.console = console;
+    }
+
+    public void printLine(String line) {
+        this.console.printLine(line);
+    }
+}
