+++ srp/src/test/java/com/codurance/srp/TransactionPrinterTest.java
@@ -1,0 +1,44 @@
+package com.codurance.srp;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InOrder;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
+
+import java.time.LocalDate;
+import java.util.Arrays;
+import java.util.List;
+
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.inOrder;
+
+@RunWith(MockitoJUnitRunner.class)
+public class TransactionPrinterTest {
+    private static final List<Transaction> TRANSACTIONS = Arrays.asList(new Transaction(LocalDate.of(2014, 4, 1), 1000), new Transaction(LocalDate.of(2014, 4, 2), -100), new Transaction(LocalDate.of(2014, 4, 10), 500));
+
+    @Mock
+    private Console console;
+    @Mock
+    private TransactionRepository transactionRepository;
+    private TransactionPrinter transactionPrinter;
+
+    @Before
+    public void setUp() {
+        transactionPrinter = new TransactionPrinter(transactionRepository, new Printer(console));
+    }
+
+    @Test
+    public void print_statement() {
+        given(transactionRepository.all()).willReturn(TRANSACTIONS);
+
+        transactionPrinter.printStatement();
+
+        InOrder inOrder = inOrder(console);
+        inOrder.verify(console).printLine("DATE | AMOUNT | BALANCE");
+        inOrder.verify(console).printLine("10/04/2014 | 500.00 | 1400.00");
+        inOrder.verify(console).printLine("02/04/2014 | -100.00 | 900.00");
+        inOrder.verify(console).printLine("01/04/2014 | 1000.00 | 1000.00");
+    }
+}
